+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-06-24T09:07:25.474Z")
+    updated: Optional[datetime] = parse_datetime("2025-06-24T11:53:50.419Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")