+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JYGK2M11V24XC682YT29VRE2"
+        "01JYGWKAZ764JPA1VS0S6R2CS4"
     )
     r"""The id of the resource"""