+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -572,5 +572,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-06-24T09:07:25.474Z")
+    updated: Optional[datetime] = parse_datetime("2025-06-24T11:53:50.419Z")
     r"""The date and time the resource was last updated"""