+++ packages/orq-rc/RELEASES.md
@@ -1829,3 +1829,13 @@ Based on:
 - [python v3.9.0-rc.4] packages/orq-rc
 ### Releases
 - [PyPI v3.9.0-rc.4] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.4 - packages/orq-rc
+
+## 2025-06-24 11:55:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.568.2 (2.634.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.9.0-rc.5] packages/orq-rc
+### Releases
+- [PyPI v3.9.0-rc.5] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.5 - packages/orq-rc